Deploy initial Gradle distribution
This commit officially opens the TurboX repository introducing
the Gradle script and the first Java class. What is in the box?

The Gradle sources have been configured as a Java library project.
Tasks have been set up to generate JAR artifacts and it's already
set up to deploy to Maven repositories.

The dependencies for libGDX and Junit have already been set up.

Finally, the propgen plugin has been set up to make the Java library
aware of compilation time variables such as version and date. A
class has been added for reading these properties and offering them
to the rest of the library via constants.