Enable Javadoc deployment from Travis CI
This gem was found on benlimmer.com's blog. It provides a system
for automatically deploying Javadocs to the website when the project
is successfully merged without no errors in master.

To do that, a private token has been generated using my account
and it has been encrypted to store it in the .travis.yml file.
When the build success, the Javadoc deployment script is executed
which, if decides that a deploy has to be made, will generate the
Javadoc, and push it to my repository overwritting the previous
documentation.