Add project information (README...)
A README file has been added with information about this repository
and a short quickstart guide. It has links to the wiki and to the
contributing guide, where further information will be found.

A CONTRIBUTING file has been added with information on the
contribution guidelines. Every user that wants to contribute to this
project should read this before doing anything.

A CHANGELOG file has been added with information on the changes
made to this project. Since this project uses GitHub Releases for
tracking changes and downloads, this file has no use.